# pup_patrol_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "pup_patrol",
        "version": "1"
    },
    "paths": {
...
        "/api/dogs": {
            "get": {
                "tags": [
                    "dog"
                ],
                "summary": "로그인된 계정에 등록된 반려견 전부 조회",
                "description": "현재 로그인 된 사람이 등록한 반려견의 정보를 data라는 키를 가진 배열로 반환",
                "operationId": "25cd015af7dc1f1288bc1c6eef66f191",
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "post": {
                "tags": [
                    "dog"
                ],
                "summary": "dog에 데이터 추가",
                "description": "dog에 데이터를 추가함",
                "operationId": "4880742d3cdf4092b2a220082d614c24",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "dogName": {
                                        "description": "반려견 이름",
                                        "type": "string",
                                        "example": "뽀삐"
                                    },
                                    "dogBreed": {
                                        "description": "반려견의 종",
                                        "type": "string",
                                        "example": "푸들"
                                    },
                                    "dogBirthDate": {
                                        "description": "반려견의 생일",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-3-2"
                                    },
                                    "dogPhotoName": {
                                        "description": "api/imageUpload 의 응답으로 나온 파일명",
                                        "type": "string",
                                        "example": "172834_뽀삐사진.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "반려견 정보가 성공적으로 추가됨"
                    }
                }
            }
        },
        "/api/dogs/{dog}": {
            "get": {
                "tags": [
                    "dog"
                ],
                "summary": "반려견 조회",
                "description": "파라미터로 받은 반려견의 id가 현재 로그인 된 사람이 등록한게 맞다면 해당 반려견 객체를 리턴함",
                "operationId": "08afb5ec931c38605d5610f8cf0fa85d",
                "parameters": [
                    {
                        "name": "dog",
                        "in": "path",
                        "description": "이용자가 등록한 반려견의 id",
                        "required": true,
                        "schema": {
                            "type": "json"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "401": {
                        "description": "본인의 반려견이 아님"
                    }
                }
            },
            "put": {
                "tags": [
                    "dog"
                ],
                "summary": "반려견 데이터 수정",
                "description": "파라미터로 넘어온 id가 로그인 한 사람이 등록한 반려견이 맞다면 반려견의 데이터를 수정함 ",
                "operationId": "1d4ac695c7350013c3edb645f5d8059b",
                "parameters": [
                    {
                        "name": "dag",
                        "in": "path",
                        "description": "반려견의 아이디",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "dogName": {
                                        "description": "반려견 이름",
                                        "type": "string",
                                        "example": "뽀삐"
                                    },
                                    "dogBreed": {
                                        "description": "반려견의 종",
                                        "type": "string",
                                        "example": "푸들"
                                    },
                                    "dogBirthDate": {
                                        "description": "반려견의 생일",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-3-2"
                                    },
                                    "dogPhotoName": {
                                        "description": "api/imageUpload 의 응답으로 나온 파일명",
                                        "type": "string",
                                        "example": "172834_뽀삐사진.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "수정이 된 경우"
                    },
                    "401": {
                        "description": "본인의 반려견이 아님"
                    }
                }
            },
            "delete": {
                "tags": [
                    "dog"
                ],
                "summary": "특정 dog 데이터 삭제",
                "description": "하나의 반려견 정보를 삭제함",
                "operationId": "b316bf22a9ac84d37b1396a412db8ff0",
                "parameters": [
                    {
                        "name": "dog",
                        "in": "path",
                        "description": "파라미터로 넘어온 id가 로그인 한 사람이 등록한 반려견이 맞다면 반려견의 데이터를 삭제함 ",
                        "required": true,
                        "schema": {
                            "type": "json"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "데이터 삭제에 성공한 경우"
                    },
                    "401": {
                        "description": "본인의 반려견이 아님"
                    }
                }
            }
        },
        "/api/imageUpload": {
            "post": {
                "tags": [
                    "image"
                ],
                "summary": "s3버킷에 이미지 추가",
                "description": "s3버킷에 이미지를 추가하고 추가된 파일명을 반환함",
                "operationId": "c9ee5974e9743b088608600d320673f9",
                "parameters": [
                    {
                        "name": "folder",
                        "in": "header",
                        "description": "이미지가 저장될 폴더명",
                        "required": true,
                        "schema": {
                            "type": "sting"
                        },
                        "example": "pup-patrol-information-image"
                    }
                ],
                "requestBody": {
                    "content": {
                        "image": {}
                    }
                },
                "responses": {
                    "201": {
                        "description": "이미지가 s3버킷에 업로드 됨"
                    },
                    "401": {
                        "description": "권한 없음"
                    }
                }
            }
        },
        "/api/imageDownload": {
            "get": {
                "tags": [
                    "image"
                ],
                "summary": "파일이름과 폴더이름을 사용해서 이미지 url을 받아옴",
                "description": "파일이름은 쿼리로 폴더이름은 헤더에서 받음",
                "operationId": "3198ec3ef6fa16cf28cf452a8f195a5c",
                "parameters": [
                    {
                        "name": "fileName",
                        "in": "query",
                        "description": "파일명",
                        "required": true,
                        "schema": {
                            "type": "sting"
                        },
                        "example": "172394_프로필사진.jpg"
                    },
                    {
                        "name": "folder",
                        "in": "header",
                        "description": "폴더명",
                        "required": true,
                        "schema": {
                            "type": "sting"
                        },
                        "example": "pup-patrol-information-image"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            }
        },
        "/api/imageDelete": {
            "delete": {
                "tags": [
                    "image"
                ],
                "summary": "s3버킷에 이미지 삭제",
                "description": "s3버킷에 이미지를 삭제함",
                "operationId": "24cfc63a652da5b87c2a3ddb250ba0bc",
                "parameters": [
                    {
                        "name": "fileName",
                        "in": "query",
                        "description": "파일명",
                        "required": true,
                        "schema": {
                            "type": "sting"
                        },
                        "example": "172394_프로필사진.jpg"
                    },
                    {
                        "name": "folder",
                        "in": "header",
                        "description": "폴더명",
                        "required": true,
                        "schema": {
                            "type": "sting"
                        },
                        "example": "pup-patrol-information-image"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "이미지가 s3버킷에서 삭제됨"
                    },
                    "401": {
                        "description": "권한 없음"
                    }
                }
            }
        },
...
    },
...
}